Changed the return signature of all functions
Functions either return a Result (Real) struct or ComplexResult struct.
This struct holds further information such as the errors in the
calculation of the routine.

+++ scify/_machine.pxd
@@ -73,4 +73,4 @@ cdef:
     double M_LNPI
     double M_EULER
 
-    double sign(double) nogil
+    int sign(double) nogil
